Use signed index type with countLabels due to cub signed-unsigned comparison warning

+++ cpp/include/raft/spatial/knn/detail/ann_kmeans_balanced.cuh
@@ -285,7 +285,7 @@ void calc_centers_and_sizes(const handle_t& handle,
 
   // Compute weight of each cluster
   raft::cluster::detail::countLabels(
-    handle, labels, temp_sizes, (uint32_t)n_rows, (uint32_t)n_clusters, workspace);
+    handle, labels, temp_sizes, (int64_t)n_rows, (int64_t)n_clusters, workspace);
 
   // Add previous sizes if necessary and cast to float
   auto counting = thrust::make_counting_iterator<int>(0);
